+++ src/main/java/roomescape/config/AdminHandlerInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.config;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
