+++ src/main/java/roomescape/controller/LogoutController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.CookieUtil;
+import roomescape.util.CookieUtil;
 
 @RestController
 @RequestMapping("/logout")
