--- src/main/java/roomescape/exception/InvalidAuthorizationException.java
@@ -1,8 +1,0 @@
-package roomescape.exception;
-
-public class InvalidAuthorizationException extends RuntimeException {
-
-    public InvalidAuthorizationException(String message) {
-        super(message);
-    }
-}
