+++ src/main/java/roomescape/util/CookieUtil.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.util;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
